feat: 拆分 messages.js 為 UI Registry 與 Error Registry
- UI 文案拆分至 shared/messaging/ui/ (10 個領域子模組)
- Error registry 拆分至 shared/messaging/errors/ (5 個子模組)
- ERROR_MESSAGES 巢狀結構由聚合層組合，shape 完全不變
- PATTERNS 交叉引用透過 import 解決
- 根目錄 messages.js 轉為 façade，既有 import 路徑不變
