+++ scripts/config/shared/index.js
@@ -4,3 +4,4 @@ export * from './highlight/index.js';
 export * from './ui/index.js';
 export * from './core/index.js';
 export * from './storage/index.js';
+export * from './messaging/index.js';
