+++ scripts/config/shared/messaging/errors/highlight.js
@@ -1,0 +1,4 @@
+export const HIGHLIGHT_ERROR_CODES = {
+  DELETE_INCOMPLETE: 'highlight_section_delete_incomplete',
+  PHASE_DELETE: 'delete_highlight_section',
+};
