+++ scripts/config/shared/messaging/errors/index.js
@@ -1,0 +1,15 @@
+import { TECHNICAL } from './technical.js';
+import { USER_MESSAGES } from './user.js';
+import { PATTERNS, DEFAULT } from './patterns.js';
+
+export const ERROR_MESSAGES = {
+  TECHNICAL,
+  USER_MESSAGES,
+  PATTERNS,
+  DEFAULT,
+};
+
+export { SECURITY_ERROR_MESSAGES } from './security.js';
+export { API_ERROR_PATTERNS } from './patterns.js';
+export { LOG_LEVELS, ERROR_TYPES } from './technical.js';
+export { HIGHLIGHT_ERROR_CODES } from './highlight.js';
