+++ scripts/config/shared/messaging/errors/security.js
@@ -1,0 +1,5 @@
+export const SECURITY_ERROR_MESSAGES = {
+  TAB_CONTEXT_REQUIRED: '拒絕訪問：此操作必須在標籤頁上下文中調用',
+  INTERNAL_ONLY: '拒絕訪問：此操作僅限擴充功能內部調用',
+  CONTENT_SCRIPT_ONLY: '拒絕訪問：僅限本擴充功能的 content script 調用',
+};
