+++ scripts/config/shared/messaging/index.js
@@ -1,0 +1,2 @@
+export { UI_MESSAGES } from './ui/index.js';
+export * from './errors/index.js';
