+++ scripts/config/shared/messaging/ui/account.js
@@ -1,0 +1,5 @@
+export const ACCOUNT = {
+  AVATAR_ALT: '使用者頭像',
+  LOCKED_LOGIN_REQUIRED: '需先登入 Google 帳號。',
+  LOCKED_COMING_SOON: '功能即將推出。',
+};
