+++ scripts/config/shared/messaging/ui/index.js
@@ -1,0 +1,25 @@
+import { AUTH, SETUP } from './auth.js';
+import { SETTINGS } from './settings.js';
+import { DATA_SOURCE } from './dataSource.js';
+import { STORAGE } from './storage.js';
+import { POPUP } from './popup.js';
+import { SIDEPANEL, HIGHLIGHTS } from './sidepanel.js';
+import { TOOLBAR } from './toolbar.js';
+import { CLOUD_SYNC } from './cloudSync.js';
+import { LOGS } from './logs.js';
+import { ACCOUNT } from './account.js';
+
+export const UI_MESSAGES = {
+  DATA_SOURCE,
+  LOGS,
+  SETTINGS,
+  AUTH,
+  ACCOUNT,
+  SETUP,
+  POPUP,
+  SIDEPANEL,
+  HIGHLIGHTS,
+  TOOLBAR,
+  STORAGE,
+  CLOUD_SYNC,
+};
