+++ scripts/config/shared/messaging/ui/logs.js
@@ -1,0 +1,5 @@
+export const LOGS = {
+  EXPORT_SUCCESS: count => `已成功匯出 ${count} 條日誌`,
+  EXPORT_FAILED_PREFIX: '匯出失敗：',
+  EXPORTING: '匯出中...',
+};
