+++ scripts/config/shared/messaging/ui/popup.js
@@ -1,0 +1,22 @@
+export const POPUP = {
+  SAVING: '儲存中...',
+  SAVE_FAILED_PREFIX: '儲存失敗：',
+  HIGHLIGHT_STARTING: '載入標註模式...',
+  HIGHLIGHT_ACTIVATED: '標註模式已啟動！',
+  HIGHLIGHT_FAILED: '啟動標註模式失敗。',
+  OPEN_NOTION_FAILED: '開啟 Notion 頁面失敗。',
+  CLEAR_CONFIRM: '確定要清除頁面上的所有標註嗎？這個操作無法復原。',
+  CLEARING: '正在清除標註...',
+  CLEAR_FAILED: '清除標註失敗。',
+  CLEAR_SUCCESS: count => `已成功清除 ${count} 條標註！`,
+  HIGHLIGHT_FAILED_PREFIX: '啟動標註失敗：',
+  PAGE_READY: '頁面已儲存，可開始標註。',
+  START_HIGHLIGHT: '開始標註',
+  DELETED_PAGE: '原頁面已刪除，請重新儲存。',
+  DELETION_PENDING: '正在確認原頁面是否已刪除，請稍後再試。',
+  SAVE_SUCCESS: '儲存成功！',
+  RECREATED: '重建成功 (原頁面已刪除)',
+  HIGHLIGHTS_UPDATED: '標註已更新',
+  UPDATED: '更新成功',
+  CREATED: '建立成功',
+};
