+++ scripts/config/shared/messaging/ui/sidepanel.js
@@ -1,0 +1,21 @@
+export const SIDEPANEL = {
+  NOT_SUPPORTED: '不支援此頁面',
+  LOAD_FAILED: '載入標註失敗',
+  SYNCING: '同步中...',
+  SYNC_SUCCESS: '同步成功',
+  SYNC_FAILED: '同步失敗',
+  OPENING: '開啟中...',
+  OPEN_SUCCESS: '開啟成功',
+  OPEN_FAILED: '開啟失敗',
+  ALL_SYNCED: '已全部同步',
+  NO_HIGHLIGHTS: '此網頁尚無標註',
+  NO_HIGHLIGHTS_SUBTITLE: '選取文字即可開始標註',
+  REMAINING_COUNT: count => `還有 ${count} 筆`,
+  PAGE_COUNT: count => `${count} 個頁面`,
+  HIGHLIGHT_COUNT: count => `${count} 個標註`,
+};
+
+export const HIGHLIGHTS = {
+  NO_NEW_TO_SYNC: '沒有新標註需要同步',
+  SYNC_SUCCESS_COUNT: count => `成功同步 ${count} 個標註`,
+};
