+++ scripts/config/shared/messaging/ui/toolbar.js
@@ -1,0 +1,6 @@
+export const TOOLBAR = {
+  SYNCING: '正在同步...',
+  SYNC_SUCCESS: '同步成功',
+  SYNC_FAILED: '同步失敗',
+  SYNC_FAILED_PREFIX: '同步失敗：',
+};
